chore: extract upgrade steps to its file
