+++ e2e/_suites/fleet/fleet.go
@@ -28,7 +28,6 @@ import (
 	"github.com/elastic/e2e-testing/internal/installer"
 	"github.com/elastic/e2e-testing/internal/kibana"
 	"github.com/elastic/e2e-testing/internal/utils"
-	"github.com/elastic/e2e-testing/pkg/downloads"
 
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
@@ -85,105 +84,6 @@ func (fts *FleetTestSuite) anStaleAgentIsDeployedToFleetWithInstaller(staleVersi
 	log.Tracef("The stale version is %s", fts.Version)
 
 	return fts.anAgentIsDeployedToFleetWithInstaller(installerType)
-}
-
-func (fts *FleetTestSuite) installCerts() error {
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
-	agentInstaller, _ := installer.Attach(fts.currentContext, fts.getDeployer(), agentService, fts.InstallerType)
-
-	err := agentInstaller.InstallCerts(fts.currentContext)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"agentVersion":      common.ElasticAgentVersion,
-			"agentStaleVersion": fts.Version,
-			"error":             err,
-			"installer":         agentInstaller,
-		}).Error("Could not install the certificates")
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"agentVersion":      common.ElasticAgentVersion,
-		"agentStaleVersion": fts.Version,
-		"error":             err,
-		"installer":         agentInstaller,
-	}).Tracef("Certs were installed")
-	return nil
-}
-
-func (fts *FleetTestSuite) anAgentIsUpgradedToVersion(desiredVersion string) error {
-	switch desiredVersion {
-	case "latest":
-		desiredVersion = common.ElasticAgentVersion
-	}
-	log.Tracef("Desired version is %s. Current version: %s", desiredVersion, fts.Version)
-
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
-
-	/*
-		// upgrading using the command is needed for stand-alone mode, only
-		agentInstaller, _ := installer.Attach(fts.currentContext, fts.getDeployer(), agentService, fts.InstallerType)
-
-		log.Tracef("Upgrading agent from %s to %s with 'upgrade' command.", desiredVersion, fts.Version)
-		return agentInstaller.Upgrade(fts.currentContext, desiredVersion)
-	*/
-
-	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
-	return fts.kibanaClient.UpgradeAgent(fts.currentContext, manifest.Hostname, desiredVersion)
-}
-
-func (fts *FleetTestSuite) agentInVersion(version string) error {
-	switch version {
-	case "latest":
-		version = downloads.GetSnapshotVersion(common.ElasticAgentVersion)
-	}
-	log.Tracef("Checking if agent is in version %s. Current version: %s", version, fts.Version)
-
-	retryCount := 0
-	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
-	exp := utils.GetExponentialBackOff(maxTimeout)
-
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
-	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
-
-	agentInVersionFn := func() error {
-		retryCount++
-
-		agent, err := fts.kibanaClient.GetAgentByHostname(fts.currentContext, manifest.Hostname)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"agent":       agent,
-				"error":       err,
-				"maxTimeout":  maxTimeout,
-				"elapsedTime": exp.GetElapsedTime(),
-				"retries":     retryCount,
-				"version":     version,
-			}).Warn("Could not get agent by hostname")
-			return err
-		}
-
-		retrievedVersion := agent.LocalMetadata.Elastic.Agent.Version
-		if isSnapshot := agent.LocalMetadata.Elastic.Agent.Snapshot; isSnapshot {
-			retrievedVersion += "-SNAPSHOT"
-		}
-
-		if retrievedVersion != version {
-			err := fmt.Errorf("version mismatch required '%s' retrieved '%s'", version, retrievedVersion)
-			log.WithFields(log.Fields{
-				"elapsedTime":      exp.GetElapsedTime(),
-				"error":            err,
-				"maxTimeout":       maxTimeout,
-				"retries":          retryCount,
-				"retrievedVersion": retrievedVersion,
-				"version":          version,
-			}).Warn("Version mismatch")
-			return err
-		}
-
-		return nil
-	}
-
-	return backoff.Retry(agentInVersionFn, exp)
 }
 
 func (fts *FleetTestSuite) agentRunPolicy(policyName string) error {
